AnimatedList bearbeitet

+++ src/my_project/model/AnimatedList.java
@@ -7,7 +7,6 @@ public class AnimatedList <T extends ListObject & AnimableList<T>> extends Graph
 
     private final List<T> list=new List();
     private T lastInList;
-    //TODO objekte der Klasse T müssen weitere Attribute; boolean isOnPointer, boolean deleted, boolean arrived, T next und T previous ; enthalten
 
     public AnimatedList(){
         list.toFirst();
@@ -72,12 +71,14 @@ public class AnimatedList <T extends ListObject & AnimableList<T>> extends Graph
     public List<T> getList(){ return list; }
 
     public void update(double dt){
-        //TODO aufruf von delete und spawn Animation nur wennbedingung erfüllt ist
         list.toFirst();
         while(list.getContent()!=null){
             T t=list.getContent();
-            t.deleteAnimation(dt);
-            t.spawnAnimation(dt);
+            if(t.isDeleted())
+                t.deleteAnimation(dt);
+            if(!t.isArrived())
+                t.spawnAnimation(dt);
+            //TODO
             list.next();
         }
     }
